Data.HABSim.Grib2.CSVParse

Description

This module provides utilities for parsing and filtering data from a wgrib2-generated CSV file. Once the CSV files is generated, it can be parsed and filtered in the following way (preferably with better error handling):

  myCsv <- readFile "/path/to/csv.csv"
  case decodeGrib myCsv of
    Left str -> error str
    Right gribLines ->
      case filterGrib 38.8977 (-77.0365) 950 gribLines of
        Nothing -> error "No entry found"
        Just wh -> print wh

Documentation

module Data.HABSim.Grib2.CSVParse.Types

decodeGrib :: ByteString -> Either String (Vector GribLine)

A helper function that primarily exists just to help type inference.

decodeGrib = decode NoHeader (where decode and NoHeader both come from Cassava).

filterGrib

Arguments

:: Double
Latitude

-> Double
Longitude

-> Int
Pressure

-> Vector GribLine
Input lines

-> Maybe GribPair
Output lines (both UGRD and VGRD)

Filter Grib lines from a Vector GribLine. If we for some reason don't have both UGRD and VGRD of our filter result, then we return Nothing. Otherwise we return a GribPair containing both.