Data.HABSim.HABSim

Documentation

module Data.HABSim.Types

sim

Arguments

:: Pitch

-> Simulation

-> Vector Int
Vector of pressures to round to from Grib file

-> HashMap Key GribLine

-> (Simulation -> Bool)
We record the line when this predicate is met

-> Writer (DList Simulation) Simulation