Data.HABSim.Internal

Documentation

m :: Double

Constants

r :: Double

Constants

er :: Double

Constants

g :: Acceleration

newVolume :: Pressure -> Volume -> Pressure -> Volume

Calculate new volume given an initial pressure and volume, and a new pressure

spRadFromVol :: Volume -> SphericalRadius

Calculate sphereical radius from volume

cAreaSp :: Double -> CrossSecArea

gas_dens :: Double -> Double -> Double -> Double

boyancy :: Double -> Double -> Double -> Double

Calculate boyancy.

drag :: Density -> Velocity -> WindMs -> CoeffDrag -> CrossSecArea -> Force

accel :: Force -> Mass -> Acceleration

force :: Mass -> Acceleration -> Force

velo :: Velocity -> Acceleration -> SimulationTime -> Velocity

displacement :: Altitude -> Velocity -> Acceleration -> SimulationTime -> Altitude

altToValues :: Altitude -> AltitudeRegionValues

altToPressure :: Altitude -> PressureDensity

roundToClosest :: (Ord a, Num a, Integral b) => a -> Vector b -> b

Given some number-like thing and a Vector of other number-like things, round the number-like thing to the closest thing in the Vector.

NOTE*: The vector must NOT be empty.

We should some day generalize this (Lens?), it seems useful.