Data.HABSim.Grib2.CSVParse.Types

Description

This module provides types for parsing CSV files generated by wgrib2. It is used (and re-exported) by CSVParse.

Documentation

data Direction

The wind direction

UGRD is North/South, VGRD is East/West.

If we're given something other than UGRD or VGRD in this field, we store the value in the Other constructor.

Constructors

UGRD

VGRD

Other String

newtype GribTime

This is used for both the referenceTime and the forecastTime. The reason it exists is solely so we can create a FromField instance on UTCTime while avoiding orphan instances.

Constructors

GribTime UTCTime

data RawGribLine

This is a (mostly-)raw gribline, right after being parsed.

The "mostly-" comes from the fact that we wrap a few things (e.g. the times into GribTime) and convert pressure into an Int.

Constructors

RawGribLine

Fields

referenceTime :: GribTime

forecastTime :: GribTime

direction :: Direction

pressure :: Int

longitude :: Double

latitude :: Double

velocity :: Double

newtype UGRDLine

A single UGRD line.

Constructors

UGRDLine RawGribLine

newtype VGRDLine

A single VGRD line.

Constructors

VGRDLine RawGribLine

data GribLine

Either a UGRDLine or a VGRDLine. This is so we can parse and ultimately return a Vector containing both UGRD and VGRD lines. We return a Vector GribLine (or rather, Cassava does) and we just know that a GribLine will either be a UGRDLine or a VGRDLine.

If the line is anything else, we return the raw line in OtherGribLine.

Constructors

UGRDGribLine UGRDLine

VGRDGribLine VGRDLine

OtherGribLine RawGribLine

data GribPair

A pair of Grib lines, where one is a UGRDLine and one is a VGRDLine. This is useful for filtering (see filterGrib below) where we want to return one of each and ensure through type safety that we don't get them confused.

Constructors

GribPair

Fields

uGrd :: UGRDLine

vGrd :: VGRDLine