Data.HABSim.Internal

Documentation

m :: Double

Constants

r :: Double

Constants

er :: Double

Constants

g :: Acceleration

newVolume :: Pressure -> Volume -> Pressure -> Volume

Calculate new volume given an initial pressure and volume, and a new pressure

spRadFromVol :: Volume -> SphericalRadius

Calculate sphereical radius from volume

cAreaSp :: Double -> CrossSecArea

gas_dens :: Double -> Double -> Double -> Double

boyancy :: Double -> Double -> Double -> Double

Calculate boyancy.

drag :: Density -> Velocity -> WindMs -> CoeffDrag -> CrossSecArea -> Force

accel :: Force -> Mass -> Acceleration

force :: Mass -> Acceleration -> Force

velo :: Velocity -> Acceleration -> SimulationTime -> Velocity

displacement :: Altitude -> Velocity -> Acceleration -> SimulationTime -> Altitude

biLinIntp :: Fractional a => a -> a -> a -> a -> a -> a -> a -> a -> a -> a -> a

altToValues :: Altitude -> AltitudeRegionValues

altToPressure :: Altitude -> PressureDensity

roundToClosest :: (Ord a, Num a, Integral b) => a -> Vector b -> b

Given some number-like thing and a Vector of other number-like things, round the number-like thing to the closest thing in the Vector.

NOTE*: The vector must NOT be empty.

We should some day generalize this (Lens?), it seems useful.

latLonBox

Arguments

:: Latitude

-> Longitude

-> Double
Resolution (units? newtype?)

-> (Latitude, Longitude, Latitude, Longitude)

Take a Longitude and Latitude, multiply it by 1 / resolution, then take the floor and ceiling of both numbers and organize them as (f,f) (f,c) (c,f) (c,c) and report this back as the bounding box.

windFromLatLon :: Latitude -> Longitude -> Int -> HashMap Key GribLine -> Maybe (WindX, WindY)

Given a Latitude, Longitude, and pressure, find the respective UGRDLine and VGRDLine if we can, from the given HashMap. Otherwise return Nothing.